requirements.txt
# ============================================================
# CYBER FORGE SCAN - Dépendances Python
# Installer avec: pip install -r requirements.txt
# ============================================================

# --- Conversion PDF ---
fpdf2>=2.7.0
python-docx>=1.0.0

# --- Test de vitesse Internet ---
speedtest-cli>=2.1.3

# --- Téléchargement YouTube ---
yt-dlp>=2024.1.0

# --- Enregistrement d'écran ---
opencv-python>=4.8.0
numpy>=1.24.0
pyautogui>=0.9.54
Pillow>=10.0.0

# --- Assistant IA (requêtes HTTP vers Ollama) ---
requests>=2.31.0
